docs: release v1.1.2

+++ mDeepFRI/tests/test_bio_utils.py
@@ -20,10 +20,11 @@ class TestBioUtils(unittest.TestCase):
 
     def test_pairwise_sqeuclidean(self):
         np.random.seed(42)
-        expected = np.array([[0, 1.01354558, 0.12442072],
-                             [1.01354558, 0, 0.99467713],
-                             [0.12442072, 0.99467713, 0]])
+        expected = np.array(
+            [[0, 1.01354558, 0.12442072], [1.01354558, 0, 0.99467713],
+             [0.12442072, 0.99467713, 0]],
+            dtype=np.float32)
 
-        matrix = np.random.rand(3, 3)
+        matrix = np.random.rand(3, 3).astype(np.float32)
         result = pairwise_sqeuclidean(matrix)
         np.allclose(result, expected)
